Struct adblock::blocker::Blocker

pub struct Blocker { /* fields omitted */ }

Implementations

impl Blocker

pub fn check(&self, request: &Request) -> BlockerResult

Decide if a network request (usually from WebRequest API) should be blocked, redirected or allowed.

pub fn check_generic_hide(&self, hostname_request: &Request) -> bool

pub fn check_parameterised(
&self,
request: &Request,
matched_rule: bool,
force_check_exceptions: bool
) -> BlockerResult

pub fn get_csp_directives(&self, _request: Request) -> Option<String>

Given a "main_frame" request, check if some content security policies should be injected in the page.

pub fn new(
network_filters: Vec<NetworkFilter>,
options: &BlockerOptions
) -> Blocker

pub fn optimize(&mut self)

If optimizations are enabled, the Blocker will be configured to automatically optimize its filters after batch updates. However, even if they are disabled, it is possible to manually call optimize(). It may be useful to have finer-grained control over optimization scheduling when frequently updating filters.

pub fn filter_exists(&self, filter: &NetworkFilter) -> bool

pub fn add_filter(&mut self, filter: NetworkFilter) -> Result<(), BlockerError>

pub fn use_tags(&mut self, tags: &[&str])

pub fn enable_tags(&mut self, tags: &[&str])

pub fn disable_tags(&mut self, tags: &[&str])

pub fn tags_enabled(&self) -> Vec<String>

pub fn use_resources(&mut self, resources: &[Resource])

pub fn add_resource(&mut self, resource: &Resource)

pub fn get_resource(&self, key: &str) -> Option<&RedirectResource>