Struct adblock::engine::Engine

pub struct Engine {
    pub blocker: Blocker,
    // some fields omitted
}

Fields
blocker: Blocker
Implementations

impl Engine

pub fn from_rules(network_filters: &[String]) -> Engine

pub fn from_rules_debug(network_filters: &[String]) -> Engine

pub fn from_rules_parametrised(
filter_rules: &[String],
load_network: bool,
load_cosmetic: bool,
debug: bool,
optimize: bool
) -> Engine

pub fn serialize(&self) -> Result<Vec<u8>, BlockerError>

pub fn deserialize(&mut self, serialized: &[u8]) -> Result<(), BlockerError>

pub fn check_network_urls(
&self,
url: &str,
source_url: &str,
request_type: &str
) -> BlockerResult

pub fn check_network_urls_with_hostnames(
&self,
url: &str,
hostname: &str,
source_hostname: &str,
request_type: &str,
third_party_request: Option<bool>
) -> BlockerResult

pub fn check_network_urls_with_hostnames_subset(
&self,
url: &str,
hostname: &str,
source_hostname: &str,
request_type: &str,
third_party_request: Option<bool>,
previously_matched_rule: bool,
force_check_exceptions: bool
) -> BlockerResult

pub fn filter_exists(&self, filter: &str) -> bool

pub fn add_filter_list<'a>(&'a mut self, filter_list: &str) -> &'a mut Engine

pub fn filter_add<'a>(&'a mut self, filter: &str) -> &'a mut Engine

pub fn with_tags<'a>(&'a mut self, tags: &[&str]) -> &'a mut Engine

pub fn tags_enable<'a>(&'a mut self, tags: &[&str])

pub fn tags_disable<'a>(&'a mut self, tags: &[&str])

pub fn tag_exists(&self, tag: &str) -> bool

pub fn with_resources<'a>(
&'a mut self,
resources: &[Resource]
) -> &'a mut Engine

pub fn resource_add<'a>(&'a mut self, resource: Resource) -> &'a mut Engine

pub fn resource_get(&self, key: &str) -> Option<RedirectResource>

pub fn hidden_class_id_selectors(
&self,
classes: &[String],
ids: &[String],
exceptions: &HashSet<String>
) -> Vec<String>

If any of the provided CSS classes or ids could cause a certain generic CSS hide rule (i.e. { display: none !important; }) to be required, this method will return a list of CSS selectors corresponding to rules referencing those classes or ids, provided that the corresponding rules are not excepted.

exceptions should be passed directly from HostnameSpecificResources.

pub fn hostname_cosmetic_resources(
&self,
hostname: &str
) -> HostnameSpecificResources

Returns a set of cosmetic filter resources required for a particular hostname. Once this has been called, all CSS ids and classes on a page should be passed to hidden_class_id_selectors to obtain any stylesheets consisting of generic rules.