Module adblock::filters::cosmetic

Tools for blocking at a page-content level, including CSS selector-based filtering and content script injection.

Structs

CosmeticFilter
Struct representing a parsed cosmetic filter rule.

CosmeticFilterMask
Boolean flags for cosmetic filter rules.

Enums

CosmeticFilterError

CosmeticFilterLocationType

Functions

get_entity_hashes_from_labels
Returns a Vec of the hashes of all segments of hostname that may match an entity-constrained rule.

get_hostname_hashes_from_labels
Returns a Vec of the hashes of all segments of hostname that may match a hostname-constrained rule.