Module adblock::lists

Structs

FilterList

Enums

FilterParseError
Unsuccessful result of parsing a single filter rule. Unused means that e.g. only network filters were requested, but a cosmetic filter was specified.

FilterType

ParsedFilter
Successful result of parsing a single filter rule

Functions

parse_filter
Parse a single filter rule

parse_filters
Parse an entire list of filters, ignoring any errors