Struct adblock::Engine

pub struct Engine { /* fields omitted */ }

The main object featured in this library. This object holds the adblocker's state, and can be queried to see if a given request should be blocked or not.

Request types

A few of Engine's methods have a field specifying a "resource type", valid examples are:

beacon

csp_report

document

font

media

object

script

stylesheet

and et cetera... See the Mozilla Web Documentation for more info.

Implementations

impl Engine

pub fn new(
network_filters: Option<Vec<String>>,
load_network: bool,
load_cosmetic: bool,
debug: bool
) -> Self

Create a new adblocking engine

pub fn check_network_urls(
&self,
url: &str,
source_url: &str,
request_type: &str
) -> BlockerResult

Check if the given url—pointing to a resource of type request_type— is blocked, assuming the request is made from the given source_url. Returns an object of type BlockerResult.

Arguments

url - The URL of the request to check

source_url - The URL from where the request is made

request_type - The resource type that the request points to

pub fn check_network_urls_with_hostnames(
&self,
url: &str,
hostname: &str,
source_hostname: &str,
request_type: &str,
third_party_request: Option<bool>
) -> BlockerResult

Check if a request should be blocked based on the given parameters.

Arguments

url - The URL of the request to check

hostname - The given url's hostname

source_hostname - The hostname of the source URL.

request_type - The resource type that the request points to

third_party_request - Is the given request to a third-party? Here, None can be given and the engine will figure it out based on the hostname and source_hostname.

pub fn check_network_urls_with_hostnames_subset(
&self,
url: &str,
hostname: &str,
source_hostname: &str,
request_type: &str,
third_party_request: Option<bool>,
previously_matched_rule: bool,
force_check_exceptions: bool
) -> BlockerResult

Check if a request should be blocked based on the given parameters.

Arguments

url - The URL of the request to check

hostname - The given url's hostname

source_hostname - The hostname of the source URL.

request_type - The resource type that the request points to

third_party_request - Is the given request to a third-party? Here, None can be given and the engine will figure it out based on the hostname and source_hostname.

previously_matched_rule - Return a match as long as there are no exceptions

force_check_exceptions - Check exceptions even if no other rule matches

pub fn serialize(&mut self) -> PyResult<Vec<u8>>

Serialize this blocking engine to bytes. They can then be deserialized using deserialize() to get the same engine again.

pub fn serialize_to_file(&mut self, file: &str) -> PyResult<()>

Serialize this blocking engine to a file. The file can then be deserialized using deserialize_from_file() to get the same engine again.

pub fn deserialize(&mut self, serialized: &[u8]) -> PyResult<()>

Deserialize a blocking engine from bytes produced with serialize().

pub fn deserialize_from_file(&mut self, file: &str) -> PyResult<()>

Deserialize a blocking engine from file produced with serialize_to_file().

pub fn add_filter_list(&mut self, filter_list: &str)

Add the contents of a block list file to the blocking engine.

pub fn filter_exists(&self, filter: &str) -> bool

Checks if the given filter exists in the blocking engine.

pub fn tags_enable(&mut self, tags: Vec<&str>)

Enable the given tags

pub fn tags_disable(&mut self, tags: Vec<&str>)

Disable the given tags

pub fn tag_exists(&self, tag: &str) -> bool

Check if the given tag exists

pub fn hostname_cosmetic_resources(
&self,
hostname: &str
) -> HostnameSpecificResources

Returns a set of cosmetic filter resources required for a particular hostname. Once this has been called, all CSS ids and classes on a page should be passed to hidden_class_id_selectors to obtain any stylesheets consisting of generic rules.

pub fn hidden_class_id_selectors(
&self,
classes: Vec<String>,
ids: Vec<String>,
exceptions: HashSet<String>
) -> PyResult<Vec<String>>

If any of the provided CSS classes or ids could cause a certain generic CSS hide rule (i.e. { display: none !important; }) to be required, this method will return a list of CSS selectors corresponding to rules referencing those classes or ids, provided that the corresponding rules are not excepted.

Exceptions should be passed directly from HostnameSpecificResources.