Struct adblock::FilterSet

pub struct FilterSet { /* fields omitted */ }

Manages a set of rules to be added to an Engine.

To be able to efficiently handle special options like $badfilter, and to allow optimizations, all rules must be available when the Engine is first created. FilterSet allows assembling a compound list from multiple different sources before compiling the rules into an Engine.

Implementations

impl FilterSet

pub fn new(debug: bool) -> Self

Creates a new FilterSet. The debug argument specifies whether or not to save information about the original raw filter rules alongside the more compact internal representation. If enabled, this information will be passed to the corresponding Engine.

pub fn add_filter_list(
&mut self,
filter_list: &str,
format: &str
) -> PyResult<()>

Adds the contents of an entire filter list to this FilterSet. Filters that cannot be parsed successfully are ignored.

The format is a string containing either "standard" (ABP/uBO-style) or "hosts".

pub fn add_filters(
&mut self,
filters: Vec<String>,
format: &str
) -> PyResult<()>

Adds a collection of filter rules to this FilterSet. Filters that cannot be parsed successfully are ignored.

The format is a string containing either "standard" (ABP/uBO-style) or "hosts".